Modified dto and some changes

+++ src/review/review.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { ReviewDocument, ReviewModel } from './review.model/review.model';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
 import { CreateReviewDto } from './dto/create-review.dto';
+import { REVIEW_NOT_FOUND, WRONG_ID } from './review.constants';
 
 @Injectable()
 export class ReviewService {
@@ -13,11 +14,26 @@ export class ReviewService {
 	}
 
 	async findByProductId(productId: string): Promise<ReviewDocument[]> {
-		return this.reviewModel.find({ productId: new Types.ObjectId(productId) }).exec();
+		if (productId.length !== 24) {
+			throw new BadRequestException(WRONG_ID);
+		}
+		const review = await this.reviewModel.find({ productId: new Types.ObjectId(productId) }).exec();
+		if (review.length === 0) {
+			throw new NotFoundException(REVIEW_NOT_FOUND);
+		}
+		return review;
 	}
 
 	async delete(id: string): Promise<ReviewDocument | null> {
-		return this.reviewModel.findByIdAndDelete(id).exec();
+		if (id.length !== 24) {
+			throw new BadRequestException(WRONG_ID);
+		}
+
+		const deletedReview = await this.reviewModel.findByIdAndDelete(id).exec();
+		if (!deletedReview) {
+			throw new NotFoundException(REVIEW_NOT_FOUND);
+		}
+		return deletedReview;
 	}
 
 	async deleteByProductId(productId: string) {
